convert CardImage to typescript
